Added currrent random directed connected graph generation code
Code is borrowed from site referenced in the README. Currently the random graphs are not producing every possible outcome but this will be changed in the future. Also produced graphs will be given an auto generated name and stored in the same folder as the source code. The way the graph files are stored needs to be changed as well.